+++ Client/src/Level/RandomNumberGenerator.h
@@ -2,5 +2,5 @@
 
 #include <cstdint>
 
-const int LehmerInt();
+const uint64_t LehmerInt();
 void SetLehmerSeed(const int seed);
